requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
opencv-python==4.1.1.26
pdf2image==1.9.0
Pillow==6.2.0
pytesseract==0.3.0
Werkzeug==0.16.0
bson==0.5.8
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
opencv-python==4.1.1.26
pdf2image==1.9.0
Pillow==6.2.0
pymongo==3.9.0
pytesseract==0.3.0
python-dateutil==2.8.1
six==1.13.0
Werkzeug==0.16.0
